ref(structure): move utils folder to src
